+++ NEWS.md
@@ -1,3 +1,13 @@
+# v0.6.3
+
+* Changed some code to match JSS paper.
+* Commented on the line where Pearson kernels are always used for factor-type variables. Should this always be the case?
+* Added control option to set intercept at a fixed value.
+* Added (hidden) options for `str()` when printing `ipriorKernel` objects.
+* Added function to find optimum Hurst coefficient for fitting FBM I-prior models.
+* Wrote vignette manual guide which details how to calculate the matrices required for the closed form estimate of `lambda`.
+* Removed the T2 statistic from the `summary()` output for now.
+
 # v0.6.2
 
 * Fix for the installation error (#26) on old R releases (prior to 3.3.0). This error was caused by the generic S3 method `sigma()` not being available from the `stats` package prior to R v3.3.0.